HAMT: Don't re-generated a seed when hash is exhausted
For a really nice implementation of HAMT, I would need the std::hash functions
to support a seed. Combining hashes with a seed (as I was doing) doesn't really
work if the whole hash collides. You just get two identical modified hashes when
there's a collission.

This commit removes the re-seeding in preparation for the solution to the hash
collision problem I have in mind.