HAMT: Finally! Handle full hash collisions properly
After the hash is exhausted when inserting, we perform a BFS to find either the
first BitmapTrie Node that's not full or the first Entry Node we can replace
with a BitmapTrie Node to fit more entries.

After the hash is exhausted when searching, we perform a DFS to find the entry
with the key. The DFS stack should not grow a lot since the height is kept at
minimum due the BFS during insertion.