support: Add some miscellaneous functions into support.h
These are necessary for the introduction of the next library.

I don't want support.h to become a kitchen-sink, but at the moment
there's no reasonable place for these functions and I don't want to
create tiny files for them because foc_libraries trades-off
modularization for the convenience of being able to copy a few files
into your project.